webapp/requirements.txt
streamlit>=1.30
anthropic>=0.39
python-pptx>=0.6.21
selenium>=4.18
webdriver-manager>=4.0
